third_party/blink/web_tests/external/wpt/editing/run/insertparagraph_1-1000-expected.txt

This is a testharness.js-based test.
Found 49 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>baz</div>" but got "foo<div>baz</div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>baz</p>" but got "foo<p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "fo[o<table><tr><td>b]ar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>fo</div><div><br></div><table><tbody><tr><td>ar</td></tr></tbody></table>" but got "fo<div><br><table><tbody><tr><td>ar</td></tr></tbody></table></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "fo[o<table><tr><td>b]ar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>fo</p><p><br></p><table><tbody><tr><td>ar</td></tr></tbody></table>" but got "fo<p><br><table><tbody><tr><td>ar</td></tr></tbody></table></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<table><tr><td>[foo<td>bar]<tr><td>baz<td>quz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div><br></div><div><br></div></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>" but got "<table><tbody><tr><td><br><br></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<table><tr><td>[foo<td>bar]<tr><td>baz<td>quz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p><br></p><p><br></p></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>" but got "<table><tbody><tr><td><br><br></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>"
[FAIL] [["insertparagraph",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<tr><td>baz<td>quz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><br></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>" but got "<table><tbody><tr><td><br><br></td><td><br></td></tr><tr><td>baz</td><td>quz</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<table><tr><td>fo[o</table>b]ar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>fo</div><div><br></div></td></tr></tbody></table>ar" but got "<table><tbody><tr><td>fo<br>ar</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<table><tr><td>fo[o</table>b]ar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>fo</p><p><br></p></td></tr></tbody></table>ar" but got "<table><tbody><tr><td>fo<br>ar</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<table><tr><td>fo[o<td>b]ar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>fo</div><div><br></div></td><td>ar</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>fo<br><br></td><td>ar</td><td>baz</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<table><tr><td>fo[o<td>b]ar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>fo</p><p><br></p></td><td>ar</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>fo<br><br></td><td>ar</td><td>baz</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "{<table><tr><td>foo</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div><div><br></div>" but got "<br><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "{<table><tr><td>foo</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><br></p><p><br></p>" but got "<br><p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<table><tr><td>[foo]</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div><br></div><div><br></div></td></tr></tbody></table>" but got "<br><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<table><tr><td>[foo]</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p><br></p><p><br></p></td></tr></tbody></table>" but got "<br><p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "[]foo" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div><div>foo</div>" but got "<div><br></div>foo"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "[]foo" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><br></p><p>foo</p>" but got "<p><br></p>foo"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "foo[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div><br></div>" but got "foo<div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "foo[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p>" but got "foo<p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<span>foo[]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span>foo</span></div><div><br></div>" but got "<span>foo</span><div><span><br></span></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<span>foo[]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span>foo</span></p><p><br></p>" but got "<span>foo</span><p><span><br></span></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "foo[]<br>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div><br></div>" but got "foo<br><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "foo[]<br>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p>" but got "foo<br><p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div>" but got "foo<div>bar</div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<p>bar</p>"
[FAIL] [["insertparagraph",""]] "<address>[]foo</address>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<address><br>foo</address>" but got "<address><br></address><address>foo</address>"
[FAIL] [["insertparagraph",""]] "<address>foo[]</address>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<address>foo<br><br></address>" but got "<address>foo</address><address><br></address>"
[FAIL] [["insertparagraph",""]] "<address>foo[]<br></address>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<address>foo<br><br></address>" but got "<address>foo<br></address><address><br></address>"
[FAIL] [["insertparagraph",""]] "<address>foo[]bar</address>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<address>foo<br>bar</address>" but got "<address>foo</address><address>bar</address>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<div>foo[]<br></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div><br></div>" but got "<div>foo<br></div><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<div>foo[]<br></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div><br></div>" but got "<div>foo<br></div><div><br></div>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo[]<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><br></dd><dd>bar</dd></dl>" but got "<dl><dt>foo</dt><dt><br></dt><dd>bar</dd></dl>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo[]<br><dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><br></dd><dd>bar</dd></dl>" but got "<dl><dt>foo<br></dt><dt><br></dt><dd>bar</dd></dl>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo<dd>bar[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar</dd><dt><br></dt></dl>" but got "<dl><dt>foo</dt><dd>bar</dd><dd><br></dd></dl>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo<dd>bar[]<br></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar</dd><dt><br></dt></dl>" but got "<dl><dt>foo</dt><dd>bar<br></dd><dd><br></dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<h1>foo[]<br></h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><div><br></div>" but got "<h1>foo<br></h1><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<h1>foo[]<br></h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><p><br></p>" but got "<h1>foo<br></h1><p><br></p>"
[FAIL] [["insertparagraph",""]] "<ol><li>foo[]<br></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li><br></li></ol>" but got "<ol><li>foo<br></li><li><br></li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo[]<br></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p>" but got "<p>foo<br></p><p><br></p>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo[]<br></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p>" but got "<p>foo<br></p><p><br></p>"
[FAIL] [["insertparagraph",""]] "<pre>[]foo</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre><br>foo</pre>" but got "<pre><br></pre><pre>foo</pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo[]</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br><br></pre>" but got "<pre>foo</pre><pre><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo[]<br></pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br><br></pre>" but got "<pre>foo<br></pre><pre><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo[]bar</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br>bar</pre>" but got "<pre>foo</pre><pre>bar</pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo[]<br><br></pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br><br><br></pre>" but got "<pre>foo</pre><pre><br><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo<br>{}<br></pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br><br><br></pre>" but got "<pre>foo<br><br></pre><pre><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo&#10;[]</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo\\n<br><br></pre>" but got "<pre>foo\\n</pre><pre><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo[]&#10;</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo<br>\\n</pre>" but got "<pre>foo\\n</pre><pre><br></pre>"
[FAIL] [["insertparagraph",""]] "<pre>foo&#10;[]&#10;</pre>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<pre>foo\\n<br>\\n</pre>" but got "<pre>foo\\n\\n</pre><pre><br></pre>"
Harness: the test ran to completion.
